Basic database implementation (#5)
The database will allow for better flexibility and extensibility in maintaining varies types of art that are requested. This features a basic implementation of adding new orders to it but you can't yet search or delete. In the future, you will also be able to export these orders as Json files for use as a receipt or import into someone else's implementation of artm.

Other features include automated category selection and hashing based on former to determine the appropriate to add. The hashing is meant to not only verify information but also allow for accurate searching and pave the way for the raffling system, see issue #1.